feat(slides): unify XML layout lint gate
(cherry picked from commit 918dc8ac4528fab1c506b79628e9fddfaffa7bdb)
